Use PAM_OLDAUTHTOK to retrieve the user's password if we're changing the password rather than PAM_AUTHTOK and save the password back into the same place.  Don't always save authentication passwords into PAM_OLDAUTHTOK; only do that if we're changing passwords.
Free the passwords retrieved from the user.  We were doing that already
when obtaining the user's password for authentication, so apparently it
is safe.

Save the new password when changing passwords into PAM_AUTHTOK if the
user enters the same password both times.

Implement PAM_PRELIM_CHECK by doing all the steps up to prompting for the
new password and changing it.  That includes authenticating the user, so
PAM_PRELIM_CHECK now may prompt.