Honor PAM_SILENT when the Kerberos library prompts for more information, passing to the application only prompts.
Avoid memory leaks in the Kerberos prompter by freeing the PAM
response strings.  We were already doing this elsewhere and the world
didn't end, so assume that it's safe for the PAM module to do this.
Also avoid memory leaks in some unusual error conditions.

Use service rather than in_tkt_service in the rest of the credential
functions.