Fix a few bugs in expired password support
In the force_pwchange case, move the password the user just typed from
AUTHTOK to OLDAUTHTOK so that the user isn't reprompted and clear
AUTHTOK so that we don't try to reset the password to the same thing.

Exit pam_chauthtok before trying to reauthenticate if we were told to
only check whether we can change passwords.