Update Autoconf krb5 portability and support krb5_get_error_string
Add support for the old Heimdal krb5_get_error_string interface.
Thanks, Chaskiel Grundman.

Add --with-krb5-include and --with-krb5-lib configure options to allow
more specific setting of paths if necessary.

If krb5-config isn't available, attempt to determine if the library
directory for the Kerberos libraries is lib32 or lib64 instead of lib
and set LDFLAGS accordingly.  Based on an idea from the CMU Autoconf
macros.