Always treat the empty password as an authentication failure
Always treat the empty password as an authentication failure rather
than passing it to the Kerberos libraries.  The Kerberos libraries
may treat it as equivalent to no password and prompt for a password
without our knowledge, leading to the user authenticating with a
different password than the one stored in the PAM stack.  This could
cause unexpected problems with some PAM configurations.  It's safer
to make the assumption that the empty password is always invalid and
reject it outside of the Kerberos libraries.  Thanks, Sanjay Sha.