Fix error handling on failure to initialize the ticket cache
Fix error handling if ticket cache initialization fails.
Authentication will still fail, but this avoids a segfault from a
double-free of the ticket cache structure.  The most common cause of
this problem was having the attempt to initialize the ticket cache
be blocked by AppArmor.  Thanks to Alex Mauer for the report.