Do not use issetugid on Solaris to check refreshing ticket caches
Do not use issetugid on Solaris to determine when to avoid refreshing
the ticket cache named in KRB5CCNAME during pam_setcred.  Instead,
compare effective and real UID and GID and permit KRB5CCNAME to be
trusted if they match.  This allows setuid screensavers on Solaris to
refresh ticket caches and makes behavior on Solaris match other
platforms.  Using issetugid is arguably safer since it protects
programs that switch users via setuid to a user other than the calling
user but still should not trust the original environment, but such
programs are rare in the PAM context and should not be calling
pam_setcred anyway unless the calling user is permitted to generally
act as the target user.  Thanks, William Yang.

issetugid is still used to determine when to call krb5_init_secure_context
and hence whether other environment variables such as KRB5_CONFIG are
trusted.