Standardize handling of use_authtok, add force_first_pass
Previous versions of this module incorrectly implemented the standard
use_authtok parameter.  use_authtok applies only to the password group
and says to use the new password stored in the PAM data rather than
prompting for a new password.  It doesn't imply anything about where
to obtain the old password, but it was implemented as requiring both
the old and new password be in the PAM stack already.  This doesn't
work when stacked with pam_cracklib.  Change use_authtok to have the
correct meaning, which means that password group configurations may
need to add use_first_pass to use_authtok to get the desired behavior.

use_first_pass and try_first_pass no longer affect how the new
password is obtained during password changes.  To use a password
obtained by a previous module, use use_authtok instead.

A new option, force_first_pass, is now supported for both the
authentication and password groups.  It tells the module to always get
the user's current password from the PAM data and fail without
prompting if it isn't already set.  This is the meaning that
use_authtok previously had for the current password.

use_authtok no longer has any meaning for the authentication stack.
Use force_first_pass instead, which does the same as use_authtok used
to do.  use_authtok will be temporarily converted to force_first_pass
in the authentication group and log a diagnostic, but this will be
removed in the future.