Change the build system to Automake and Libtool
pam-krb5 is now built using Automake and Libtool to bring it more in
line with other software packages.  This means that it now relies on
Libtool to know how to generate a loadable module rather than
hand-configured linker rules.  This may improve portability on some
platforms and may hurt it on other platforms.

Remove install-sh, config.guess, and config.sub from the Git repository
since autoreconf is now responsible for adding them to the distribution.