Prompt for the new password even for ignored users
In the second pass through the password group, prompt for the new
password and store it in the PAM data even if the user is being
ignored.  This is required to allow this module to be stacked with
another module that uses use_authtok.  Without this behavior, the
second module won't be able to work for any ignored user since it will
see no saved password and use_authtok will reject the password change.