Use pam_syslog or pam_vsyslog for all logging
Change the syslog message format to match the Linux PAM standard
by using pam_syslog or pam_vsyslog for all logging.  Replace all
uses of static buffers in logging with asprintf or vasprintf.
Change how the authenticating user is presented to use a prefix
of (user %s) instead of %s:.