Do not prompt for a password with try_pkinit and MIT
Do not prompt for a password when try_pkinit is set and the module is
built against MIT Kerberos.  This fixes a spurious password prompt but
introduces a bug where the user's password is not saved in the PAM
data if the authentication falls back to password when PKINIT fails.
This requires more work to fix and will be addressed in a subsequent
release.  Thanks to Бранко Мајић for the report.