Supress expired password notification with {force,use}_first_pass
Suppress the notice that the password is being changed because it's
expired if force_first_pass or use_first_pass is set in the password
stack, indicating that it's stacked with another module that's also
doing password changes.  This is arguable, but without this change the
notification message of why the password is being changed shows up
confusingly in the middle of the password change interaction.  Based
on a patch by William Yang.