Switch to the pam-util-provided pam_args struct
Rename the internal pam_args struct to pam_config and set it as a
member of the pam-util-provided parent pam_args struct.  Modify all
users of the old pam_args struct to know about the new one instead.

Several of these changes are temporary hacks that will be backed out
after more of the conversion is done.