Reuse the Kerberos context where possible
If we're creating a new pam-krb5 context, use the Kerberos context
that we already created as part of PAM setup where we can.  Make sure
that we don't free it twice.  Handle some additional bookkeeping to
coordinate the idea of the current authenticated user in our various
data structures.  Simplify some of the cleanup code.