Handle krb5_get_init_creds_opt_set_pkinit in portability layer
Rather than using an #ifdef in the code, handle the different argument
count in portable/krb5.h and just call the current version in the
main code.  Modify the Autoconf macro to check for the obsolete
version rather than the current version and reverse the sense of the
defined C preprocessor symbol.