Check output line by line and add support for wildcards
Rather than checking all output from the module as one string, check
it line-by-line.  Add support for a %* wildcard that matches anything
following it in the line when matching output.

This required changing pam_output in the fake PAM library to return
a vector, including some build system hackery to make the vector
library available to fake PAM and changes to the other test cases.