Add new tests for password change of expired accounts
This is a bit of a mess because the default behavior uses the library
to prompt, which means we need separate scripts for MIT and Heimdal
(and they're likely to need to change with future versions).

Also check defer_pwchange and fail_pwchange, but only on MIT since we
have to be able to disable library prompting for password change,
which we can't do on Heimdal.