Revert setting OLDAUTHTOK on defer_pwchange, improve authz
Don't set PAM_OLDAUTHTOK on defer_pwchange.  All the auth tokens are
cleared when pam_authenticate exits, so it's pointless.  Improve the
documentation in auth.c about why we do what we do with expired
passwords.

Do the .k5login check even if the user has authenticated with an
expired password and defer_pwchange is set.  We don't want to return
success if the authentication principal isn't in .k5login.