Add TAP utility function to generate a testing krb5.conf
Move the code to generate a testing krb5.conf file into a utility
function and call it instead of duplicating the code.  Use an
atexit handler to free the environment variable and delete the
file.  Move the file into tests/tmp instead of the main tests
directory, following the new pattern towards using a tmp directory
for generated files.