Refactor the prompting code
Unwind a lot of the code in pamk5_password_auth to use functions and
be easier to read.  Separate the retrieval of authtoks and the prompting
for passwords into separate functions, and separate a single pass through
the possible authentication methods into a separate function.  Add const
to the password retrieved from the PAM data.