Add new no_prompt option to suppress PAM module password prompting
Add a new option no_prompt, which suppresses the PAM module's prompt
for the user's password and defers all prompting to the Kerberos
library.  This allows the Kerberos library to have complete control of
the prompting process, which may be desireable if authentication
mechanisms other than password are in use.  Be aware that, with this
option set, the PAM module has no control over the contents of the
prompt and cannot store the user's password in the PAM data.  Based on
a patch by Yair Yarom.

The test suite will only pass for Heimdal currently.  This will be
fixed in a subsequent commit.