Add regex matching support to the fakepam testing scripts
Allow both the prompt and an output line to be a regular expression,
marked by starting and ending it with a slash (/).  Skip such tests
if regular expression support is not available in the C library.

The response to a prompt now begins with the last | on the line,
allowing | to be used as a regex metacharacter.