Track output and priority separately, fix regexes for output
Rather than adding the priority and a space to the beginning of each
output line, use a more complex output struct and capture the priority
and output line separately.

Take advantage of that in the script testing framework to properly
support regex matching for output lines, since we now recognize that
the regex is the entirety of the output line since the priority is
not appended to the front.