First pass of coding style cleanup for anonymous FAST
Reformat and rearchitect the FAST code to try to make it cleaner
and more readable.  This is the first of several passes.

Always use krb5_get_init_creds_opt_set_fast_ccache_name, since that's
the most widely supported interface and since it's not at all clear
from the API documentation whether the version that just takes the
ticket cache requires that the ticket cache stay open.