Add preliminary support for Kerberos trace logging
Add preliminary support for Kerberos trace logging via a trace option
that enables trace logging if supported by the underlying Kerberos
library.  The option takes as an argument the file name to which to
log trace output.  This option does not yet work with any released
version of Kerberos, but may work with the next release of MIT
Kerberos.