Update to new mechanism of valgrind testing
Use the C TAP Harness support for valgrind testing and mark the
tests that can be run under valgrind.  Fix problems in one module
that knew too much about the contents of the test temporary
directory.  Exclude the tests that use the pkinit.so module, since
they cause valgrind to go into an infinite loop.