Support PAM_DATA_SILENT with pam_end
Properly support calling pam_end with PAM_DATA_SILENT by not deleting
the underlying ticket cache.  This flag is used when the application
is closing the PAM session after a fork to free memory resources, but
doesn't intend to free resources external to the process because
another process may still depend on them.  Thanks to Andrew G. Morgan
for the report.

Fixes #21